+++ src/main/java/es/udc/fi/tfg/controller/TagRestController.java
@@ -78,27 +78,37 @@ public class TagRestController {
 	
 	@PostMapping(value = "/admin/tags")
 	public ResponseEntity<Tag> createTag(@RequestBody Tag tag) throws EntityNotCreatableException {
-		tagService.createTag(tag);
-		return new ResponseEntity<Tag>(HttpStatus.CREATED);
+		if (tagService.existsTag(tag)){
+			return new ResponseEntity<Tag>(HttpStatus.CONFLICT);	
+		}
+		else {
+			tagService.createTag(tag);
+			return new ResponseEntity<Tag>(HttpStatus.CREATED);			
+		}
 	}
 	
 	@DeleteMapping("/admin/tags/{id}")
 	public ResponseEntity<String> deleteTag(@PathVariable int id) {
-		int rows;
-		try{
-			rows = tagService.deleteTag(id);
+		if (tagService.getTag(id)==null) {
+			return new ResponseEntity<String>(HttpStatus.NOT_FOUND);
 		}
-		catch(EntityNotRemovableException e){
-			return new ResponseEntity<String>(HttpStatus.LOCKED);
-		}
-		catch(Exception e){
-			return new ResponseEntity<String>("No ha podido eliminarse el Tag: " + id, HttpStatus.INTERNAL_SERVER_ERROR);
-		}
-		if (rows < 1) {
-			return new ResponseEntity<String>("No ha podido eliminarse el Tag "+id, HttpStatus.INTERNAL_SERVER_ERROR);
-		}
-		else{
-			return new ResponseEntity<String>(HttpStatus.OK);
+		else {
+			int rows;
+			try{
+				rows = tagService.deleteTag(id);
+			}
+			catch(EntityNotRemovableException e){
+				return new ResponseEntity<String>(HttpStatus.LOCKED);
+			}
+			catch(Exception e){
+				return new ResponseEntity<String>("No ha podido eliminarse el Tag: " + id, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+			if (rows < 1) {
+				return new ResponseEntity<String>("No ha podido eliminarse el Tag "+id, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+			else{
+				return new ResponseEntity<String>(HttpStatus.OK);
+			}			
 		}
 	}
 
@@ -106,6 +116,9 @@ public class TagRestController {
 	public ResponseEntity<String> updateTag(@PathVariable int id, @RequestBody Tag tag) throws EntityNotUpdatableException {
 		if (tag.getId()!=id) {
 			return new ResponseEntity<String>("Los ids no coinciden"+id, HttpStatus.BAD_REQUEST);
+		}
+		else if (tagService.getTag(id)==null) {
+			return new ResponseEntity<String>(HttpStatus.NOT_FOUND);
 		}
 		else {
 			int rows = tagService.updateTag(id, tag);
